mysql.js file split in CRUD form files and create quiz sql request completed
