+++ .gitignore
@@ -1,1 +1,2 @@
 node_modules/
+controlers/js/sqlParameters.js
