+++ controlers/js/sqlDelete.js
@@ -1,0 +1,4 @@
+const mysql = require('mysql');
+const connectionParameters = require('./sqlParameters');
+
+// TODO
