+++ views/pages/faq.ejs
@@ -23,14 +23,18 @@
 
         <input type="checkbox" id="q1">
         <label class="label" for="q1">Comment partager mon quiz ? </label>
-        <p class="etat">Chaque quiz créé aura son propre code pin unique ainsi qu'une adresse unique. A la fin de votre quiz, vous n'aurez plus qu'à partager ce code pin à votre entourage pour qu'il puisse profiter de votre quiz.  </p>
+        <p class="etat">
+            Chaque quiz créé aura son propre code pin unique ainsi qu'une adresse unique. A la fin de votre quiz, vous n'aurez plus qu'à partager ce code pin à votre entourage pour qu'il puisse profiter de votre quiz. 
+        </p>
         </div>
 
         <div class="faq">
 
         <input type="checkbox" id="q2">
         <label class="label" for="q2">Mes quizs sont-ils publics ? </label>
-        <p class="etat">Tous les quizs qui sont publiés sur la plateforme deviennent public dès qu'ils auront été validé par notre équipe. Les quizs seront disponibles directement sur la plateforme dans l'onglet jouer mais vous pourez aussi rechercher un quiz par mots-clés, par thématique ou par popularité.</p>
+        <p class="etat">
+            Tous les quizs qui sont publiés sur la plateforme deviennent public dès qu'ils auront été validé par notre équipe. Les quizs seront disponibles directement sur la plateforme dans l'onglet jouer mais vous pourez aussi rechercher un quiz par mots-clés, par thématique ou par popularité.
+        </p>
 
 
         </div>
@@ -39,7 +43,10 @@
 
         <input type="checkbox" id="q3">
         <label class="label" for="q3">Un quiz avec + de 10 questions, c'est possible ? </label>
-        <p class="etat">Malheuresement, nous ne premettons que la création d'un quiz comportant 10 questions et 4 réponses. Si vous avez besoin de poser de nouvelles questions, rien ne vous empêche de créer un nouveau quiz qui sera la deuxième partie. </p>
+        <p class="etat">
+            Malheureusement, nous ne premettons que la création d'un quiz comportant 10 questions et 4 réponses. Si vous avez besoin de poser de nouvelles questions, rien ne vous empêche de créer un nouveau quiz qui sera la deuxième partie. 
+        </p>
+
 
 
         </div>
@@ -48,7 +55,9 @@
 
         <input type="checkbox" id="q4">
         <label class="label" for="q4">Est-ce que des enfants peuvent jouer ? </label>
-        <p class="etat">Cette plateforme a été créée pour tous les publics. L'interface graphique a été pensée pour être simple d'utilisation. Chacun quiz a fait l'objet d'une sérieuse correction pour garantir un contenu intégre et respectant les bonnes moeurs. </p>
+        <p class="etat">
+            Cette plateforme a été créée pour tous les publics. L'interface graphique a été pensée pour être simple d'utilisation. Chacun quiz a fait l'objet d'une sérieuse correction pour garantir un contenu intégre et respectant les bonnes moeurs.
+        </p>
 
 
         </div>
@@ -57,7 +66,9 @@
 
         <input type="checkbox" id="q5">
         <label class="label" for="q5">Est-ce que plusieurs réponses sont possibles ? </label>
-        <p class="etat">Plusieurs réponses sont effectivement possibles pour chaque question. Nous vous conseillons de lire attentivement les énoncés avant de réponse. Vous avez en général suffisament de temps pour répondre aux questions. </p>
+        <p class="etat">
+            Il n'y a qu'une et une seule bonne réponse par question. Nous vous conseillons de lire attentivement les énoncés avant de réponse. Vous avez en général suffisament de temps pour répondre aux questions.
+        </p>
 
 
         </div>
@@ -66,7 +77,9 @@
 
         <input type="checkbox" id="q6">
         <label class="label" for="q6">Puis-je commencer la rédaction d'un quiz et y revenir plus tard ? </label>
-        <p class="etat">Si vous vous arrêtez de rédiger votre quiz pour n'importe quelle raison, vous pouvez cliquer sur l'onglet "sauvegarder". Un numéro d'identification s'affichera. Faites bien attention de le conserver. Il vous permettra de reprendre là où vous vous êtes arrété.  </p>
+        <p class="etat">
+            Si vous vous arrêtez de rédiger votre quiz pour n'importe quelle raison, vous pouvez cliquer sur l'onglet "sauvegarder". Un numéro d'identification s'affichera. Faites bien attention de le conserver. Il vous permettra de reprendre là où vous vous êtes arrété. 
+        </p>
 
         </div>
 
@@ -74,7 +87,9 @@
 
         <input type="checkbox" id="q7">
         <label class="label" for="q7">Qu'est-ce que je gagne si je réponds bien à toutes les réponses. </label>
-        <p class="etat">Aucun lot ne sont à gagner. Chaque créateur de quiz sera tout de même libre de vous indiquer un petit mot d'encouragement. /p>
+        <p class="etat">
+            Aucun lot ne sont à gagner. Chaque créateur de quiz sera tout de même libre de vous indiquer un petit mot d'encouragement.
+        </p>
 
         </div>
 
@@ -82,7 +97,9 @@
 
         <input type="checkbox" id="q8">
         <label class="label" for="q8">Peut-on inventer n'importe quel quiz ? </label>
-        <p class="etat">Les thématiques de quizs sont libres mais elle corresponde généralement à une logique d'apprentissage. Idéalement, le quiz devra permettre au joueur de tester ses connaissances et améliorer sa culture générale. </p>
+        <p class="etat">
+            Les thématiques de quizs sont libres mais elle corresponde généralement à une logique d'apprentissage. Idéalement, le quiz devra permettre au joueur de tester ses connaissances et améliorer sa culture générale.
+        </p>
 
 
         </div>
