+++ views/pages/questionspage.ejs
@@ -82,7 +82,7 @@
                 } else if ( (score > 0) && (score < 5)) {
                     number[0].textContent = 'Une petite révision s\'impose';
                 } else if ( (score > 5) && (score < 10)) {
-                    number[0].textContent = 'Jolie score';
+                    number[0].textContent = 'Joli score';
                 } else if ( (score === 10)) {
                     number[0].textContent = 'Votre maitrise du sujet est indiscutable !';
                 } else {
